// src/app/shared/interfaces/category.ts
import { CustomFields } from './custom-fields';

export interface Category {
    id: number;
    type: 'shop'|'blog';
    categoryName: string;
    slug: string;
    path: string;
    image: string|null;
    items: number;
    customFields: CustomFields;
    parents?: Category[];
    children?: Category[];
}